Add event listener
Event listeners were added to get the onclick Javascript functions
out of the HTML file.

Two event listeners were added one for the form and one for the
button. Now, the 'Enter' button or clicking the button will submit
the form.

The decimal place was set to 2 instead of 4 when the total amount
is printed.